requirements.txt
backoff==1.6.0
beautifulsoup4==4.6.1
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
idna==2.7
lxml==4.2.4
praw==6.0.0
prawcore==1.0.0
python-dotenv==0.9.1
requests==2.19.1
update-checker==0.16
urllib3==1.23
